3863 Drop usePipelineValidation
