+++ src/pageEditor/tabs/editTab/EditTab.tsx
@@ -42,14 +42,12 @@ import cx from "classnames";
 import { FormState } from "@/pageEditor/pageEditorTypes";
 import useReportTraceError from "./useReportTraceError";
 import FoundationNodeConfigPanel from "./FoundationNodeConfigPanel";
-import usePipelineValidation from "@/pageEditor/validation/usePipelineValidation";
 
 const EditTab: React.FC<{
   eventKey: string;
 }> = ({ eventKey }) => {
   useExtensionTrace();
   useReportTraceError();
-  usePipelineValidation();
 
   const { values } = useFormikContext<FormState>();
